Move `DEPENDENT_KEYS_*` constants
